some view do not exists

+++ tools/init_SIGA2025VVC.sh
@@ -9,7 +9,11 @@ extract_SIGA2025VVC() {
     for frame in $(seq 1 $2); do
         rm -rf $ROOT/frame$frame/images
         mkdir -p $ROOT/frame$frame/images
+        echo "Organizing frame $frame"
         for view in $(seq 1 $3); do
+            if [ ! -e $ROOT/images/$(printf "%02d" $((view-1))) ] || [ ! -e $ROOT/masks/$(printf "%02d" $((view-1))) ]; then
+                continue
+            fi
             mv $ROOT/images/$(printf "%02d" $((view-1)))/$(printf "%06d" $((frame-1))).jpg $ROOT/frame$frame/images/$(printf "%02d" $((view-1))).jpg
             mv $ROOT/masks/$(printf "%02d" $((view-1)))/$(printf "%06d" $((frame-1))).png $ROOT/frame$frame/images/$(printf "%02d" $((view-1)))_mask.png
         done
